@@ -11,25 +11,30 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.9",
-    "@angular/cdk": "^13.3.7",
-    "@angular/common": "~13.3.9",
-    "@angular/compiler": "~13.3.9",
-    "@angular/core": "~13.3.9",
-    "@angular/forms": "~13.3.9",
-    "@angular/material": "^13.3.7",
-    "@angular/platform-browser": "~13.3.9",
-    "@angular/platform-browser-dynamic": "~13.3.9",
-    "@angular/router": "~13.3.9",
+    "@angular/animations": "^14.1.1",
+    "@angular/cdk": "^14.1.1",
+    "@angular/common": "^14.1.1",
+    "@angular/compiler": "^14.1.1",
+    "@angular/core": "^14.1.1",
+    "@angular/forms": "^14.1.1",
+    "@angular/material": "^14.1.1",
+    "@angular/platform-browser": "^14.1.1",
+    "@angular/platform-browser-dynamic": "^14.1.1",
+    "@angular/router": "^14.1.1",
+    "@fortawesome/angular-fontawesome": "^0.11.0",
+    "@fortawesome/fontawesome-svg-core": "~1.2.36",
+    "@fortawesome/free-brands-svg-icons": "^5.15.4",
+    "@fortawesome/free-regular-svg-icons": "^5.15.4",
+    "@fortawesome/free-solid-svg-icons": "^5.15.4",
     "rxjs": "~6.6.0",
     "swiper": "^8.2.2",
     "tslib": "^2.3.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.3.6",
-    "@angular/cli": "~13.3.6",
-    "@angular/compiler-cli": "~13.3.9",
+    "@angular-devkit/build-angular": "^14.1.1",
+    "@angular/cli": "^14.1.1",
+    "@angular/compiler-cli": "^14.1.1",
     "@types/jasmine": "~3.8.0",
     "@types/node": "^12.11.1",
     "jasmine-core": "~3.8.0",
